+++ src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/ShopDB
+spring.datasource.username=UsershopDB
+spring.datasource.password=UsershopDB
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=validate
+
+spring.liquibase.change-log=classpath:liquibase/changelog-master.yml
